backend/requirements.txt
flask
flask-cors
replicate
python-dotenv
requests
google-generativeai
mediapipe
gunicorn
opencv-python
numpy
Pillow
torch
torchvision
transformers